Create InitialDataEval.R
Wrote this code to analyse the data just to get some visual pictures of it (boxplot or scatterplot) and print some summaries.
Narrowed down from all values for ones that I think would realistically make sense to examine. 

Output <- number of occurences of 9,98,99,999 in all variables chosen to consider more. 
analyseBabies <- the correlations to consider 

This removes all instances of 9, 99, 999, and 98 so its definitely skewed a bit but it was a good way to check initial evaluations on a lot of columns to choose the ones to look into more.